apps/backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Sandstone backend API"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.135.1",
    "pydantic-settings>=2.13.1",
    "python-docx>=1.2.0",
    "python-multipart>=0.0.17",
    "uvicorn>=0.41.0",
]

[tool.pytest.ini_options]
pythonpath = ["."]

[dependency-groups]
dev = [
    "httpx>=0.28.0",
    "pytest>=9.0.2",
]
